refactor(api): split generated auth schema
Keep Better Auth generated tables in auth-schema.ts and re-export them from the app schema. This keeps the auth generator from rewriting app-owned tables.
