+++ apps/api/package.json
@@ -20,7 +20,7 @@
 		"dev:dashboard": "bun run --cwd ../dashboard build && bun run dev:local",
 		"deploy": "bun run --cwd ../dashboard build && wrangler deploy",
 		"typegen": "wrangler types",
-		"auth:generate": "infisical run --env=prod --path=/api -- bun x @better-auth/cli generate --yes --config ./better-auth.config.ts --output ./src/db/schema.ts",
+		"auth:generate": "infisical run --env=prod --path=/api -- bun x @better-auth/cli generate --yes --config ./better-auth.config.ts --output ./src/db/auth-schema.ts",
 		"db:generate": "drizzle-kit generate",
 		"db:drop": "drizzle-kit drop",
 		"db:push:local": "drizzle-kit push",
