+++ apps/api/src/db/auth-schema.ts
@@ -1,0 +1,343 @@
+import { relations } from 'drizzle-orm';
+import {
+	boolean,
+	index,
+	jsonb,
+	pgTable,
+	text,
+	timestamp,
+} from 'drizzle-orm/pg-core';
+
+export const user = pgTable('user', {
+	id: text('id').primaryKey(),
+	name: text('name').notNull(),
+	email: text('email').notNull().unique(),
+	emailVerified: boolean('email_verified').default(false).notNull(),
+	image: text('image'),
+	createdAt: timestamp('created_at').defaultNow().notNull(),
+	updatedAt: timestamp('updated_at')
+		.defaultNow()
+		.$onUpdate(() => /* @__PURE__ */ new Date())
+		.notNull(),
+});
+
+export const session = pgTable(
+	'session',
+	{
+		id: text('id').primaryKey(),
+		expiresAt: timestamp('expires_at').notNull(),
+		token: text('token').notNull().unique(),
+		createdAt: timestamp('created_at').defaultNow().notNull(),
+		updatedAt: timestamp('updated_at')
+			.$onUpdate(() => /* @__PURE__ */ new Date())
+			.notNull(),
+		ipAddress: text('ip_address'),
+		userAgent: text('user_agent'),
+		activeOrganizationId: text('active_organization_id'),
+		userId: text('user_id')
+			.notNull()
+			.references(() => user.id, { onDelete: 'cascade' }),
+	},
+	(table) => [index('session_userId_idx').on(table.userId)],
+);
+
+export const organization = pgTable('organization', {
+	id: text('id').primaryKey(),
+	name: text('name').notNull(),
+	slug: text('slug').notNull().unique(),
+	logo: text('logo'),
+	metadata: jsonb('metadata'),
+	createdAt: timestamp('created_at').defaultNow().notNull(),
+	updatedAt: timestamp('updated_at')
+		.defaultNow()
+		.$onUpdate(() => /* @__PURE__ */ new Date()),
+});
+
+export const member = pgTable(
+	'member',
+	{
+		id: text('id').primaryKey(),
+		organizationId: text('organization_id')
+			.notNull()
+			.references(() => organization.id, { onDelete: 'cascade' }),
+		userId: text('user_id')
+			.notNull()
+			.references(() => user.id, { onDelete: 'cascade' }),
+		role: text('role').default('member').notNull(),
+		createdAt: timestamp('created_at').defaultNow().notNull(),
+	},
+	(table) => [
+		index('member_organization_id_idx').on(table.organizationId),
+		index('member_user_id_idx').on(table.userId),
+	],
+);
+
+export const invitation = pgTable(
+	'invitation',
+	{
+		id: text('id').primaryKey(),
+		organizationId: text('organization_id')
+			.notNull()
+			.references(() => organization.id, { onDelete: 'cascade' }),
+		email: text('email').notNull(),
+		role: text('role').notNull(),
+		status: text('status').default('pending').notNull(),
+		expiresAt: timestamp('expires_at').notNull(),
+		inviterId: text('inviter_id')
+			.notNull()
+			.references(() => user.id, { onDelete: 'cascade' }),
+		createdAt: timestamp('created_at').defaultNow().notNull(),
+	},
+	(table) => [
+		index('invitation_organization_id_idx').on(table.organizationId),
+		index('invitation_email_idx').on(table.email),
+		index('invitation_inviter_id_idx').on(table.inviterId),
+	],
+);
+
+export const account = pgTable(
+	'account',
+	{
+		id: text('id').primaryKey(),
+		accountId: text('account_id').notNull(),
+		providerId: text('provider_id').notNull(),
+		userId: text('user_id')
+			.notNull()
+			.references(() => user.id, { onDelete: 'cascade' }),
+		accessToken: text('access_token'),
+		refreshToken: text('refresh_token'),
+		idToken: text('id_token'),
+		accessTokenExpiresAt: timestamp('access_token_expires_at'),
+		refreshTokenExpiresAt: timestamp('refresh_token_expires_at'),
+		scope: text('scope'),
+		password: text('password'),
+		createdAt: timestamp('created_at').defaultNow().notNull(),
+		updatedAt: timestamp('updated_at')
+			.$onUpdate(() => /* @__PURE__ */ new Date())
+			.notNull(),
+	},
+	(table) => [index('account_userId_idx').on(table.userId)],
+);
+
+export const verification = pgTable(
+	'verification',
+	{
+		id: text('id').primaryKey(),
+		identifier: text('identifier').notNull(),
+		value: text('value').notNull(),
+		expiresAt: timestamp('expires_at').notNull(),
+		createdAt: timestamp('created_at').defaultNow().notNull(),
+		updatedAt: timestamp('updated_at')
+			.defaultNow()
+			.$onUpdate(() => /* @__PURE__ */ new Date())
+			.notNull(),
+	},
+	(table) => [index('verification_identifier_idx').on(table.identifier)],
+);
+
+export const jwks = pgTable('jwks', {
+	id: text('id').primaryKey(),
+	publicKey: text('public_key').notNull(),
+	privateKey: text('private_key').notNull(),
+	createdAt: timestamp('created_at').notNull(),
+	expiresAt: timestamp('expires_at'),
+});
+
+export const oauthClient = pgTable('oauth_client', {
+	id: text('id').primaryKey(),
+	clientId: text('client_id').notNull().unique(),
+	clientSecret: text('client_secret'),
+	disabled: boolean('disabled').default(false),
+	skipConsent: boolean('skip_consent'),
+	enableEndSession: boolean('enable_end_session'),
+	scopes: text('scopes').array(),
+	userId: text('user_id').references(() => user.id, { onDelete: 'cascade' }),
+	createdAt: timestamp('created_at'),
+	updatedAt: timestamp('updated_at'),
+	name: text('name'),
+	uri: text('uri'),
+	icon: text('icon'),
+	contacts: text('contacts').array(),
+	tos: text('tos'),
+	policy: text('policy'),
+	softwareId: text('software_id'),
+	softwareVersion: text('software_version'),
+	softwareStatement: text('software_statement'),
+	redirectUris: text('redirect_uris').array().notNull(),
+	postLogoutRedirectUris: text('post_logout_redirect_uris').array(),
+	tokenEndpointAuthMethod: text('token_endpoint_auth_method'),
+	grantTypes: text('grant_types').array(),
+	responseTypes: text('response_types').array(),
+	public: boolean('public'),
+	type: text('type'),
+	requirePKCE: boolean('require_pkce'),
+	referenceId: text('reference_id'),
+	metadata: jsonb('metadata'),
+});
+
+export const oauthRefreshToken = pgTable('oauth_refresh_token', {
+	id: text('id').primaryKey(),
+	token: text('token').notNull(),
+	clientId: text('client_id')
+		.notNull()
+		.references(() => oauthClient.clientId, { onDelete: 'cascade' }),
+	sessionId: text('session_id').references(() => session.id, {
+		onDelete: 'set null',
+	}),
+	userId: text('user_id')
+		.notNull()
+		.references(() => user.id, { onDelete: 'cascade' }),
+	referenceId: text('reference_id'),
+	expiresAt: timestamp('expires_at'),
+	createdAt: timestamp('created_at'),
+	revoked: timestamp('revoked'),
+	authTime: timestamp('auth_time'),
+	scopes: text('scopes').array().notNull(),
+});
+
+export const oauthAccessToken = pgTable('oauth_access_token', {
+	id: text('id').primaryKey(),
+	token: text('token').unique(),
+	clientId: text('client_id')
+		.notNull()
+		.references(() => oauthClient.clientId, { onDelete: 'cascade' }),
+	sessionId: text('session_id').references(() => session.id, {
+		onDelete: 'set null',
+	}),
+	userId: text('user_id').references(() => user.id, { onDelete: 'cascade' }),
+	referenceId: text('reference_id'),
+	refreshId: text('refresh_id').references(() => oauthRefreshToken.id, {
+		onDelete: 'cascade',
+	}),
+	expiresAt: timestamp('expires_at'),
+	createdAt: timestamp('created_at'),
+	scopes: text('scopes').array().notNull(),
+});
+
+export const oauthConsent = pgTable('oauth_consent', {
+	id: text('id').primaryKey(),
+	clientId: text('client_id')
+		.notNull()
+		.references(() => oauthClient.clientId, { onDelete: 'cascade' }),
+	userId: text('user_id').references(() => user.id, { onDelete: 'cascade' }),
+	referenceId: text('reference_id'),
+	scopes: text('scopes').array().notNull(),
+	createdAt: timestamp('created_at'),
+	updatedAt: timestamp('updated_at'),
+});
+
+export const userRelations = relations(user, ({ many }) => ({
+	sessions: many(session),
+	accounts: many(account),
+	memberships: many(member),
+	invitationsSent: many(invitation),
+	oauthClients: many(oauthClient),
+	oauthRefreshTokens: many(oauthRefreshToken),
+	oauthAccessTokens: many(oauthAccessToken),
+	oauthConsents: many(oauthConsent),
+}));
+
+export const sessionRelations = relations(session, ({ one, many }) => ({
+	user: one(user, {
+		fields: [session.userId],
+		references: [user.id],
+	}),
+	oauthRefreshTokens: many(oauthRefreshToken),
+	oauthAccessTokens: many(oauthAccessToken),
+}));
+
+export const organizationRelations = relations(organization, ({ many }) => ({
+	members: many(member),
+	invitations: many(invitation),
+}));
+
+export const memberRelations = relations(member, ({ one }) => ({
+	organization: one(organization, {
+		fields: [member.organizationId],
+		references: [organization.id],
+	}),
+	user: one(user, {
+		fields: [member.userId],
+		references: [user.id],
+	}),
+}));
+
+export const invitationRelations = relations(invitation, ({ one }) => ({
+	organization: one(organization, {
+		fields: [invitation.organizationId],
+		references: [organization.id],
+	}),
+	inviter: one(user, {
+		fields: [invitation.inviterId],
+		references: [user.id],
+	}),
+}));
+
+export const accountRelations = relations(account, ({ one }) => ({
+	user: one(user, {
+		fields: [account.userId],
+		references: [user.id],
+	}),
+}));
+
+export const oauthClientRelations = relations(oauthClient, ({ one, many }) => ({
+	user: one(user, {
+		fields: [oauthClient.userId],
+		references: [user.id],
+	}),
+	oauthRefreshTokens: many(oauthRefreshToken),
+	oauthAccessTokens: many(oauthAccessToken),
+	oauthConsents: many(oauthConsent),
+}));
+
+export const oauthRefreshTokenRelations = relations(
+	oauthRefreshToken,
+	({ one, many }) => ({
+		oauthClient: one(oauthClient, {
+			fields: [oauthRefreshToken.clientId],
+			references: [oauthClient.clientId],
+		}),
+		session: one(session, {
+			fields: [oauthRefreshToken.sessionId],
+			references: [session.id],
+		}),
+		user: one(user, {
+			fields: [oauthRefreshToken.userId],
+			references: [user.id],
+		}),
+		oauthAccessTokens: many(oauthAccessToken),
+	}),
+);
+
+export const oauthAccessTokenRelations = relations(
+	oauthAccessToken,
+	({ one }) => ({
+		oauthClient: one(oauthClient, {
+			fields: [oauthAccessToken.clientId],
+			references: [oauthClient.clientId],
+		}),
+		session: one(session, {
+			fields: [oauthAccessToken.sessionId],
+			references: [session.id],
+		}),
+		user: one(user, {
+			fields: [oauthAccessToken.userId],
+			references: [user.id],
+		}),
+		oauthRefreshToken: one(oauthRefreshToken, {
+			fields: [oauthAccessToken.refreshId],
+			references: [oauthRefreshToken.id],
+		}),
+	}),
+);
+
+export const oauthConsentRelations = relations(oauthConsent, ({ one }) => ({
+	oauthClient: one(oauthClient, {
+		fields: [oauthConsent.clientId],
+		references: [oauthClient.clientId],
+	}),
+	user: one(user, {
+		fields: [oauthConsent.userId],
+		references: [user.id],
+	}),
+}));
